app/requirements.txt
fastapi
uvicorn
python-dotenv
asyncpg
psycopg2-binary
sqlalchemy
alembic
pydantic-settings
PyJWT
python-multipart
